Fix badge causing mobile overflow issues
