+++ src/components/NewBadge.js
@@ -1,9 +1,16 @@
 import React from "react";
-import {
-  Badge,
-
-} from "@material-ui/core";
+import {Badge} from "@material-ui/core";
+import {makeStyles} from "@material-ui/styles";
 import moment from "moment";
+
+const useStyles = makeStyles((theme) => ({
+  badge: {
+    display: "flex",
+    position: "relative",
+    flexShrink: 0,
+    verticalAlign: "middle",
+  },
+}));
 
 const daysOld = (fromDate) => {
   return moment().diff(fromDate, "days");
@@ -11,9 +18,11 @@ const daysOld = (fromDate) => {
 
 const NewBadge = ({fromDate, children, youngerThan=10}) => {
   const hideBadge = daysOld(fromDate) > youngerThan;
+  const classes = useStyles();
 
   return (
     <Badge anchorOrigin={{vertical: "top", horizontal: "left"}}
+      classes={{root: classes.badge}}
       overlap="rectangle" badgeContent="NEW" color="error"
       invisible={hideBadge}>{children}</Badge>
   );
